+++ packages/sources/react/src/components/structure/VtmnSkeleton/index.ts
@@ -1,0 +1,2 @@
+export { default as VtmnSkeleton } from './VtmnSkeleton';
+export * from './VtmnSkeleton';