+++ packages/sources/vue/src/components/forms/VtmnSelect/types.ts
@@ -1,4 +1,5 @@
 export interface VtmnSelectOption {
   label: string | number | boolean | null;
   value: string | number | boolean | null;
+  disabled?: boolean;
 }