+++ packages/sources/vue/src/components/index.ts
@@ -11,6 +11,7 @@ import VtmnChip from './selection-controls/VtmnChip';
 import VtmnRadioButton from './selection-controls/VtmnRadioButton';
 import VtmnToggle from './selection-controls/VtmnToggle';
 import VtmnDivider from './structure/VtmnDivider';
+import VtmnSkeleton from './structure/VtmnSkeleton';
 
 export {
   VtmnButton,
@@ -26,4 +27,5 @@ export {
   VtmnTag,
   VtmnToggle,
   VtmnDivider,
+  VtmnSkeleton,
 };